fixing the similarity matrix gen

+++ SED2013/src/main/java/org/openimaj/mediaeval/evaluation/solr/SED2013Index.java
@@ -275,7 +275,7 @@ public class SED2013Index {
 		}
 
 		public static IndexedPhoto fromDoc(SolrDocument doc){
-			long photoIndex = (long) doc.get("index");
+			long photoIndex = (long)(Long) doc.get("index");
 			Photo p = PhotoUtils.createPhoto(doc);
 			return new IndexedPhoto(photoIndex, p);
 		}
